requirements.txt
flask
flask-cors
sentence-transformers
pinecone
requests
gunicorn
waitress
python-dotenv